LLM_LIAR/requirements.txt
# requirements.txt
openai==1.9.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
nltk==3.8.1
python-dotenv==1.0.0